fix(dir 75): optimize pages loading (#767)
* Refactor layout and add skeleton loaders

Reorganized layout components to separate filters and content panels. Introduced skeleton loaders for better user experience during content loading. Updated CSS to improve alignment and consistency across components.

* Remove unused imports and improve CSS consistency

Unused imports were removed across several files to clean up the codebase. Additionally, CSS files were standardized by removing unnecessary "px" units where zero values were specified to improve consistency and readability.

* Refactor events page layout and optimize component structure

Restructured the events page by introducing a new layout. Simplified component responsibilities, added loading states, and modularized CSS for better maintainability. Removed redundant props to streamline data flow.

* Remove unused `isLoggedIn` and fix `aggregatedEventsData` scope

The `isLoggedIn` variable was removed from the `Events` function as it was unused, decluttering the code. Additionally, the `aggregatedEventsData` variable's declaration scope was fixed to ensure proper assignment and maintain cleaner code structure.

* Update skeleton loader to use dynamic items count

Replaced hardcoded value with `INITIAL_ITEMS_PER_PAGE` constant for better configurability and alignment with pagination settings. This makes the skeleton loader dynamically adjust to match the initial page load size.
