+++ app/events/(events-page)/@contributors/loading.module.css
@@ -1,0 +1,31 @@
+.root {
+    height: 100%;
+    flex: 1;
+    gap: 14px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 20px;
+}
+
+.header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, 36px);
+    row-gap: 8px;
+    column-gap: 8px;
+}
+
+.content {
+    display: flex;
+    flex-direction: row;
+}
+
+.content > span {
+    width: 100%;
+}
