+++ app/events/(events-page)/@events/loading.module.css
@@ -1,0 +1,26 @@
+.root {
+    height: 100%;
+    flex: 1;
+    gap: 14px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    padding: 32px 20px 20px;
+}
+
+.header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.grid {
+    display: grid;
+    grid-template-columns: repeat(6, 285px);
+    row-gap: 24px;
+    column-gap: 16px;
+}
+
+.grid span {
+    height: 290px;
+}
