--- app/events/page.module.css
@@ -1,43 +1,0 @@
-.eventsPage {
-  width: 100%;
-}
-
-.huskyBannerContainer {
-  width: 100%;
-  padding: 0 20px 30px 20px;
-}
-
-.contributorsSection {
-  width: 100%;
-}
-
-.container {
-  width: 100%;
-  height: auto;
-}
-
-.shedule__hdr__btn {
-  display: flex;
-  gap: 1rem;
-  align-items: center;
-  justify-content: center;
-  margin-top: 1rem;
-}
-
-@media (min-width: 768px) {
-  .shedule__hdr__btn {
-    justify-content: flex-end;
-    margin-top: 0;
-  }
-
-  .container {
-    padding: 20px;
-  }
-}
-
-@media (min-width: 1024px) {
-  .eventsPage {
-    padding: 20px;
-    background: #BFEAFF;
-  }
-}
