+++ app/members/(members-page)/@content/loading.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { ContentPanelSkeletonLoader } from '@/components/core/dashboard-pages-layout/ContentPanelSkeletonLoader';
+
+const Loading = () => {
+  return (
+    <ContentPanelSkeletonLoader />
+  );
+};
+
+export default Loading;
