+++ app/members/(members-page)/@content/page.module.css
@@ -7,7 +7,7 @@
 }
 
 .members__left {
-  width: 0px;
+  width: 0;
 }
 
 .members__right {
@@ -44,7 +44,7 @@
   .members__right {
     display: flex;
     justify-content: center;
-    padding: 30px 0px 40px 0px;
+    padding: 30px 0 40px 0;
   }
 
   .members__left__filterweb {
@@ -58,8 +58,8 @@
   }
 
   .members__right__membersList {
-    margin-top: 0px;
-    padding: 0px 16px;
+    margin-top: 0;
+    padding: 0 16px;
   }
 
   .members__left {
