+++ app/projects/(projects-page)/@content/page.module.css
@@ -8,15 +8,26 @@
 }
 
 .project__filter {
-  width: 0px;
+  width: 0;
 }
 
 .project__cn {
-  flex: 1;
-  background: #f1f5f9;
+  /*flex: 1;*/
+  /*background: #f1f5f9;*/
+  width: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 .project__cn__toolbar {
+  /*height: 72px;*/
+  /*position: sticky;*/
+  /*z-index: 4;*/
+  /*top: 80px;*/
+  /*padding: 16px;*/
+  /*border-bottom: 1px solid #e2e8f0;*/
+  /*background: #fff;*/
+
   height: 72px;
   position: sticky;
   z-index: 4;
@@ -45,20 +56,29 @@
   }
 
   .project__cn {
-    padding: 22px 32px 32px 32px;
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
+    /*padding: 22px 32px 32px 32px;*/
+    /*display: flex;*/
+    /*flex-direction: column;*/
+    /*gap: 24px;*/
+    gap: 14px;
   }
 
   .project__cn__toolbar {
-    height: 40px;
+    /*height: 40px;*/
+    /*height: 60px;*/
+    /*padding: 16px;*/
+    /*padding: 0px;*/
+    /*border: none;*/
+    /*background: #f1f5f9;*/
+    /*padding-top: 10px;*/
+
     height: 60px;
-    padding: 16px;
-    padding: 0px;
+    padding: 10px 40px;
     border: none;
     background: #f1f5f9;
-    padding-top: 10px;
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
   }
 
   .project__filter__web {
