+++ components/core/dashboard-pages-layout/index.ts
@@ -1,0 +1,2 @@
+export * from './FiltersPanelSkeletonLoader';
+export * from './DashboardPagesLayout';
